refactor: clean up Markdown component imports and remove unused functions
