+++ frontend/shared/components/markdown/streamdown-render.tsx
@@ -22,21 +22,23 @@ import {
   MarkdownImage,
   MarkdownLink,
   MarkdownParagraph,
-  MarkdownHTMLArticle,
-  MarkdownHTMLAside,
-  MarkdownHTMLDetails,
-  MarkdownHTMLDiv,
-  MarkdownHTMLMain,
-  MarkdownHTMLSection,
-  MarkdownHTMLSpan,
-  MarkdownHTMLSummary,
   MarkdownArtifactActionsContext,
-  MarkdownHTMLMarkdownRendererContext,
   MarkdownStrong,
   ThinkingHeading,
   type MarkdownArtifactActions,
   type MarkdownImageActions,
 } from "./streamdown-components";
+import {
+  MarkdownHTMLArticle,
+  MarkdownHTMLAside,
+  MarkdownHTMLDetails,
+  MarkdownHTMLDiv,
+  MarkdownHTMLMain,
+  MarkdownHTMLMarkdownRendererContext,
+  MarkdownHTMLSection,
+  MarkdownHTMLSpan,
+  MarkdownHTMLSummary,
+} from "./streamdown-html";
 import {
   normalizeContent,
   normalizeCurrencyDollars,
